requirements.txt
python-igraph
goatools
cairocffi